Split JuliaInterface/src/gap_macros.c
... into calls.c and convert.c, and compile those files separately
instead of #including into src/JuliaInterface.c
